+++ CHANGELOG.md
@@ -10,8 +10,7 @@
 ### Features
 
 + 合并了 v1 中的一些中间件
-+ 新增了中间件： cors、swagger、logger、rest、graphql
-+ 将一些 ctx 的扩展操作移到 x 包中（预发）
++ 新增了中间件： cors、swagger、logger、rest、graphql、(ws)WebSocket、xrouter
 + 增加独立的 box 仓库存放扩展中间件
 
 ### Breaking Changes